fixing warnings, refs #3

+++ src/microsim/traffic_lights/MSRailSignal.cpp
@@ -102,7 +102,7 @@ MSRailSignal::init(NLDetectorBuilder&) {
     }
     updateCurrentPhase();
     setTrafficLightSignals(MSNet::getInstance()->getCurrentTimeStep());
-    myNumLinks = myLinks.size();
+    myNumLinks = (int)myLinks.size();
 }
